requirements.txt
numpy==1.21.4
transforms3d==0.3.1
sapien==1.1.1
open3d==0.13.0
rospy==1.15.13
rosbag==1.15.13
matplotlib==3.5.0
tqdm==4.62.3
scikit-learn==1.0.1
PyYAML==6.0
opencv-python==4.5.4.60
Pillow==8.4.0
pandas==1.3.4
yourdfpy==0.0.52
h5py==3.6.0
pybullet==3.2.0
pymeshlab==2021.10
urdfpy==0.0.22